MICROSOFT-ENTRA-AGENT-ID/examples/basic-agent-auth/requirements.txt
# Microsoft Entra Agent ID - Basic Authentication Example
# Updated: January 19, 2026

# Azure Identity SDK
azure-identity>=1.15.0
azure-core>=1.29.0

# Azure Storage SDK (for examples)
azure-storage-blob>=12.19.0

# Token handling
PyJWT>=2.8.0
cryptography>=41.0.0

# HTTP client
requests>=2.31.0

# Environment configuration
python-dotenv>=1.0.0

# Logging and monitoring
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
